+++ EmailSender/src/main/java/com/pritam/email/sender/EmailSenderApplication.java
@@ -1,4 +1,4 @@
-package com.pritam.EmailSender;
+package com.pritam.email.sender;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
