+++ EmailSender/src/test/java/com/pritam/email/sender/EmailSenderApplicationTests.java
@@ -1,4 +1,4 @@
-package com.pritam.EmailSender;
+package com.pritam.email.sender;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
